requirements.txt
requests>=2.28.0
python-dotenv>=1.0.0
pydantic>=2.0.0
flask>=3.0.0
pytest>=7.0.0
httpx>=0.25.0
